Some BuildCraft support added. Pretty much done, though the Coil acts a bit weirdly with it.

+++ common/net/lomeli/magiks/api/magiks/IMagiks.java
@@ -10,11 +10,11 @@ public interface IMagiks
 
     void addToMistLevel(int value);
 
-    int getHeatLevel();
+    float getHeatLevel();
 
-    void setHeatLevel(int temp);
+    void setHeatLevel(float temp);
 
-    void addToHeatLevel(int temp);
+    void addToHeatLevel(float temp);
 
     int getMaxMistLevel();
